Cleanup: consolidate duplicated ingredients, products code
